MB-6580 Fix large file regression
Reviewed-on: http://review.couchbase.org/20679
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>

+++ src/os.c
@@ -156,7 +156,7 @@ static void couch_close(couch_file_handle handle)
 static off_t couch_goto_eof(couch_file_handle handle)
 {
     int fd = handle_to_fd(handle);
-    int rv = lseek(fd, 0, SEEK_END);
+    off_t rv = lseek(fd, 0, SEEK_END);
     if(rv < 0) {
         save_errno();
     }
